+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - REMIX-4739: Add support for specific icons for the different logic node categories
 - REMIX-4576: Added icon support for prims tab
 - Small improvements for Remix Logic: added delete button, hid unused ui, stage manager refresh filtering
+- REMIX-4719: Added Stage Manager Remix Logic Graph Interaction Action Plugin, And Logic Graph Filter
 
 ### Changed
 - Update hdremix and omni_core_materials to ext-822f7b6-main