+++ source/extensions/lightspeed.trex.app.resources/docs/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.26.0]
+### Added
+- Added LogicGraphWidgetPlugin Widget to the Stage Manager via schema configuration
 
 ## [1.25.0]
 ### Added