+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.6.0]
+### Added
+- Added `LogicGraphWidgetPlugin` to the interactions plugins
+
 ## [2.5.0]
 ### Added
 - Added `ignore_property_prefix_events` to filter stage manager refresh for "ui:" prefixed properties by default